requirements.txt
numpy
asteroid-filterbanks >=0.4
einops >=0.6.0
huggingface_hub >= 0.13.0
lightning >= 2.0.4
omegaconf >=2.1,<3.0
pyannote.core >= 5.0.0
pyannote.database >= 5.0.1
pyannote.metrics >= 3.2
pyannote.pipeline >= 2.3   # 2.4
pytorch_metric_learning >= 2.1.0
rich >= 12.0.0
semver >= 3.0.0
soundfile >= 0.12.1
speechbrain >= 0.5.14
tensorboardX >= 2.6
torch >= 2.0.1
torch_audiomentations >= 0.11.0
torchaudio >= 2.0.0
torchmetrics >= 0.11.0
https://github.com/pyannote/pyannote-audio/archive/refs/heads/develop.zip #pyannote.audio
torchvision
numba
tqdm
more-itertools
tiktoken==0.3.3
scikit-learn
pydiarization
# pyaudio==0.2.11
pydub
# openai-whisper
yellowbrick
tensorflow
stt